[UI] Add alias for mojo version of `MenuSourceType`
This CL removes `MenuSourceType` enum from ui_base_types.h.
Added alias for mojo version until all clients are migrated.
Once every client is migrated, alias will be removed.

`MenuSourceType` is moved from ui_base_types.mojom to it's own
mojom due to circular dependency between `ui/base/mojom` and `ui/base`.
`MenuSourceType` will be moved back to ui_base_types.mojom once
the circular dependency is gone.

Changed a few ui::MenuSourceType::MENU_SOURCE_* to ui::MENU_SOURCE_* to
enable build. These will be changed to use mojo version later.

AX-Relnotes: N/A
Bug: 355460976
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5803393
Auto-Submit: Md Hasibul Hasan <hasibulhasan873@gmail.com>
Commit-Queue: Md Hasibul Hasan <hasibulhasan873@gmail.com>
Cr-Commit-Position: refs/heads/main@{#1367879}

+++ content/browser/android/selection/selection_popup_controller.cc
@@ -281,7 +281,8 @@ bool SelectionPopupController::ShowSelectionMenu(
       params.selection_rect.y(), params.selection_rect.right(),
       params.selection_rect.bottom(), handle_height, params.is_editable,
       is_password_type, jselected_text, params.selection_start_offset,
-      can_select_all, can_edit_richly, should_suggest, params.source_type,
+      can_select_all, can_edit_richly, should_suggest,
+      static_cast<int>(params.source_type),
       render_frame_host->GetJavaRenderFrameHost());
   return true;
 }
